refactor(backend): separate tower and warp
