+++ crates/stellation-backend-tower/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "stellation-backend"
+name = "stellation-backend-tower"
 version = "0.1.4"
 edition = "2021"
 rust-version = "1.66"
@@ -15,26 +15,14 @@ homepage = "https://github.com/futursolo/stellation"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
+hyper = { version = "0.14.23", features = ["runtime", "server", "http1"] }
+tower = { version = "0.4", features = ["util"] }
+tokio = { version = "1" }
 futures = { version = "0.3", default-features = false, features = ["std"] }
-typed-builder = "0.11.0"
-serde = { version = "1", features = ["derive"] }
-thiserror = "1"
-thread_local = "1.1.4"
-once_cell = "1.17.0"
-lol_html = "0.3.2"
-tracing = { version = "0.1.37" }
-serde_urlencoded = "0.7.1"
+yew = { version = "0.20", features = ["ssr"] }
+warp = { version = "0.3.3", default-features = false }
 
 # Stellation Components
+stellation-backend-warp = { version = "0.1.4", path = "../stellation-backend-warp" }
+stellation-backend = { version = "0.1.4", path = "../stellation-backend" }
 stellation-bridge = { version = "0.1.4", path = "../stellation-bridge", features = ["resolvable"] }
-stellation-core = { version = "0.1.4", path = "../stellation-core" }
-
-# Yew / Component Related
-yew = { version = "0.20", features = ["ssr"] }
-bounce = { version = "0.6", features = ["helmet", "ssr"] }
-yew-router = "0.17"
-
-
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "documenting"]
