+++ crates/stellation-backend-warp/src/frontend.rs
@@ -1,20 +1,16 @@
-use std::borrow::Cow;
 use std::ops::Deref;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::{fmt, str};
 
-use bounce::helmet::HelmetTag;
-use lol_html::{doc_comments, element, rewrite_str, Settings};
 use rust_embed::{EmbeddedFile, RustEmbed};
+use stellation_backend::utils::ThreadLocalLazy;
 use tokio::fs;
 use warp::filters::fs::File;
 use warp::filters::BoxedFilter;
 use warp::path::Tail;
 use warp::reply::{with_header, Response};
 use warp::{Filter, Rejection, Reply};
-
-use crate::utils::ThreadLocalLazy;
 
 type GetFileFn = Box<dyn Send + Fn(&str) -> Option<EmbeddedFile>>;
 
@@ -117,82 +113,14 @@ pub(crate) enum IndexHtml {
 }
 
 impl IndexHtml {
-    async fn read_content(&self) -> Cow<'_, str> {
+    pub async fn read_content(&self) -> Arc<str> {
         match self {
             IndexHtml::Path(p) => fs::read_to_string(&p)
                 .await
-                .map(Cow::from)
+                .map(Arc::from)
                 .expect("TODO: implement failure."),
 
-            IndexHtml::Embedded(ref s) => s.as_ref().into(),
+            IndexHtml::Embedded(ref s) => s.clone(),
         }
-    }
-
-    pub async fn render<I, H, B>(&self, tags: I, head_s: H, body_s: B) -> String
-    where
-        I: IntoIterator<Item = HelmetTag>,
-        H: Into<String>,
-        B: AsRef<str>,
-    {
-        let mut head_s = head_s.into();
-        let body_s = body_s.as_ref();
-
-        let mut html_tag = None;
-        let mut body_tag = None;
-
-        for tag in tags.into_iter() {
-            match tag {
-                HelmetTag::Html { .. } => {
-                    html_tag = Some(tag);
-                }
-                HelmetTag::Body { .. } => {
-                    body_tag = Some(tag);
-                }
-                _ => {
-                    let _ = tag.write_static(&mut head_s);
-                }
-            }
-        }
-
-        let index_html_s = self.read_content().await;
-
-        rewrite_str(
-            &index_html_s,
-            Settings {
-                element_content_handlers: vec![
-                    element!("html", |h| {
-                        if let Some(HelmetTag::Html { attrs }) = html_tag.take() {
-                            for (k, v) in attrs {
-                                h.set_attribute(k.as_ref(), v.as_ref())?;
-                            }
-                        }
-
-                        Ok(())
-                    }),
-                    element!("body", |h| {
-                        if let Some(HelmetTag::Body { attrs }) = body_tag.take() {
-                            for (k, v) in attrs {
-                                h.set_attribute(k.as_ref(), v.as_ref())?;
-                            }
-                        }
-
-                        Ok(())
-                    }),
-                ],
-
-                document_content_handlers: vec![doc_comments!(|c| {
-                    if c.text() == "%STELLATION_HEAD%" {
-                        c.replace(&head_s, lol_html::html_content::ContentType::Html);
-                    }
-                    if c.text() == "%STELLATION_BODY%" {
-                        c.replace(body_s, lol_html::html_content::ContentType::Html);
-                    }
-
-                    Ok(())
-                })],
-                ..Default::default()
-            },
-        )
-        .expect("failed to render html")
     }
 }
