+++ crates/stellation-backend-warp/src/utils.rs
@@ -1,0 +1,13 @@
+//! Warp utilities.
+
+/// Creates a random string.
+pub(crate) fn random_str() -> String {
+    use rand::distributions::Alphanumeric;
+    use rand::Rng;
+
+    rand::thread_rng()
+        .sample_iter(&Alphanumeric)
+        .take(7)
+        .map(char::from)
+        .collect()
+}
