+++ examples/fullstack/server/src/app.rs
@@ -1,9 +1,12 @@
 use example_fullstack_view::Main;
-use stellation_backend::ServerAppProps;
+use stellation_backend::{Request, ServerAppProps};
 use yew::prelude::*;
 
 #[function_component]
-pub fn ServerApp(_props: &ServerAppProps<()>) -> Html {
+pub fn ServerApp<REQ>(_props: &ServerAppProps<(), REQ>) -> Html
+where
+    REQ: Request,
+{
     html! {
         <Main />
     }
