+++ examples/fullstack/server/src/main.rs
@@ -2,8 +2,8 @@
 #![deny(missing_debug_implementations)]
 
 use example_fullstack_api::create_bridge;
-use stellation_backend::Endpoint;
 use stellation_backend_cli::Cli;
+use stellation_backend_tower::TowerEndpoint;
 
 mod app;
 use app::ServerApp;
@@ -17,10 +17,11 @@ struct Frontend;
 async fn main() -> anyhow::Result<()> {
     stellation_backend_cli::trace::init_default("STELLATION_APP_SERVER_LOG");
 
-    let endpoint = Endpoint::<ServerApp>::new().with_bridge(create_bridge());
+    let endpoint = TowerEndpoint::<ServerApp<_>>::new().with_bridge(create_bridge());
 
     #[cfg(stellation_embedded_frontend)]
-    let endpoint = endpoint.with_frontend(stellation_backend::Frontend::new_embedded::<Frontend>());
+    let endpoint =
+        endpoint.with_frontend(stellation_backend_tower::Frontend::new_embedded::<Frontend>());
 
     Cli::builder().endpoint(endpoint).build().run().await?;
